Issue 1401: Text color for white-theme buttons.
Currently the buttons do not have an explicit text color. This works if
the device default theme has a proper color for the text by default, but
fails if the default color is not readable.

This commit makes the color explicit and therefore makes it the same on
all devices.