Refresh to Travis CI configuration.
This commit updates the Travis Continuous Integration configuration to
make it work again.

I chose a different approach to installing the SDK, that avoid depending
on the order in which things are returned by the tool and breaks each
time something changes (like a new release). Instead use static links to
the SDK components that we need.

Since the continuous build is not reliable enough (or at least I am not
sure it is), also only notify myself of failures.