Remove Unescaper class.
The class had a dubious copyright and can be easily implemented using
the Android APIs.

This commit also adds a test project that I used to create a couple of
simple unit tests for the newly added code.

The project can be configured in Eclipse by importing the tests
directory as a new project, as well as with ant, running "ant test". See
the README in the tests directory for more details.