Switch to original note editor.
The multimedia note editor needs a bit more work before it is ready for
release. This commit switches back to the original note editor, so that
we can release a bug-fix version of the application without having to
wait until the multimedia note editor is ready.

It also removed the RECORD_AUDIO permission, since it is not needed if
the multimedia note editor is not used.